Adopt new project structure
