+++ example/chunks/golang/Dockerfile
