+++ example/tests/golang.yaml
