+++ pkg/dazzle/build.go
@@ -2,17 +2,13 @@ package dazzle
 
 import (
 	"context"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"os"
-	"path/filepath"
 	"sort"
-	"strings"
 
-	"github.com/bmatcuk/doublestar"
 	"github.com/containerd/console"
 	"github.com/containerd/containerd/errdefs"
 	clog "github.com/containerd/containerd/log"
@@ -22,124 +18,19 @@ import (
 	"github.com/csweichel/dazzle/pkg/test/buildkit"
 	"github.com/docker/distribution/reference"
 	"github.com/mattn/go-isatty"
-	"github.com/minio/highwayhash"
 	"github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/session/auth/authprovider"
 	"github.com/moby/buildkit/util/progress/progressui"
 	"github.com/opencontainers/go-digest"
 	ociv1 "github.com/opencontainers/image-spec/specs-go/v1"
-	ignore "github.com/sabhiram/go-gitignore"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/sync/errgroup"
-	"gopkg.in/yaml.v2"
 )
 
 var (
 	hashKey = []byte{0x03, 0x40, 0xf3, 0xc8, 0x94, 0x7c, 0xad, 0x78, 0x75, 0x14, 0x0f, 0x4c, 0x4a, 0xf7, 0xc6, 0x2b, 0x43, 0x13, 0x1d, 0xc2, 0xa8, 0xc7, 0xfc, 0x46, 0x28, 0xf0, 0x68, 0x5e, 0x36, 0x9a, 0x3b, 0x0b}
 )
-
-// ProjectConfig is the structure of a project's dazzle.yaml
-type ProjectConfig struct {
-	ChunkIgnore  []string `yaml:"ignore"`
-	chunkIgnores *ignore.GitIgnore
-}
-
-// Project is a dazzle build project
-type Project struct {
-	Base   ProjectChunk
-	Chunks []ProjectChunk
-	Config ProjectConfig
-}
-
-// ProjectChunk represents a layer chunk in a project
-type ProjectChunk struct {
-	Name        string
-	Dockerfile  []byte
-	ContextPath string
-	Tests       []*test.Spec
-
-	cachedHash string
-}
-
-// LoadFromDir loads a dazzle project from disk
-func LoadFromDir(dir string) (*Project, error) {
-	var (
-		cfg   ProjectConfig
-		cfgfn = filepath.Join(dir, "dazzle.yaml")
-	)
-	if fd, err := os.Open(cfgfn); err == nil {
-		log.WithField("filename", cfgfn).Debug("loading dazzle config")
-		err = yaml.NewDecoder(fd).Decode(&cfg)
-		fd.Close()
-		if err != nil {
-			return nil, fmt.Errorf("cannot load config from %s: %w", cfgfn, err)
-		}
-
-		cfg.chunkIgnores, err = ignore.CompileIgnoreLines(cfg.ChunkIgnore...)
-		if err != nil {
-			return nil, fmt.Errorf("cannot load config from %s: %w", cfgfn, err)
-		}
-	}
-
-	base, err := loadChunk(dir, "_base")
-	if err != nil {
-		return nil, err
-	}
-
-	res := &Project{
-		Base: *base,
-	}
-	chds, err := ioutil.ReadDir(dir)
-	if err != nil {
-		return nil, err
-	}
-
-	res.Chunks = make([]ProjectChunk, 0, len(chds))
-	for _, chd := range chds {
-		if cfg.chunkIgnores != nil && cfg.chunkIgnores.MatchesPath(chd.Name()) {
-			continue
-		}
-		if strings.HasPrefix(chd.Name(), "_") || strings.HasPrefix(chd.Name(), ".") {
-			continue
-		}
-		if !chd.IsDir() {
-			continue
-		}
-		chnk, err := loadChunk(dir, chd.Name())
-		if err != nil {
-			return nil, err
-		}
-		res.Chunks = append(res.Chunks, *chnk)
-		log.WithField("name", chnk.Name).Info("added chunk to project")
-	}
-
-	return res, nil
-}
-
-func loadChunk(base, name string) (*ProjectChunk, error) {
-	dir := filepath.Join(base, name)
-	res := &ProjectChunk{
-		Name:        name,
-		ContextPath: dir,
-	}
-
-	var err error
-	res.Dockerfile, err = ioutil.ReadFile(filepath.Join(dir, "Dockerfile"))
-	if err != nil {
-		return nil, err
-	}
-
-	tf, err := ioutil.ReadFile(filepath.Join(base, fmt.Sprintf("%s-tests.yaml", name)))
-	if err != nil && !os.IsNotExist(err) {
-		return nil, fmt.Errorf("%s: cannot read tests.yaml: %w", dir, err)
-	}
-	err = yaml.UnmarshalStrict(tf, &res.Tests)
-	if err != nil {
-		return nil, fmt.Errorf("%s: cannot read tests.yaml: %w", dir, err)
-	}
-	return res, nil
-}
 
 type buildOpts struct {
 	CacheRef           reference.Named
@@ -535,80 +426,6 @@ func getImageMetadata(ctx context.Context, ref reference.Reference, resolver rem
 	return
 }
 
-func (p *ProjectChunk) hash(baseref string) (res string, err error) {
-	if p.cachedHash != "" {
-		return p.cachedHash, nil
-	}
-
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("cannot compute hash: %w", err)
-		}
-	}()
-
-	hash, err := highwayhash.New(hashKey)
-	if err != nil {
-		return
-	}
-
-	err = p.manifest(baseref, hash)
-	if err != nil {
-		return
-	}
-
-	res = hex.EncodeToString(hash.Sum(nil))
-	p.cachedHash = res
-	return
-}
-
-func (p *ProjectChunk) manifest(baseref string, out io.Writer) (err error) {
-	sources, err := doublestar.Glob(filepath.Join(p.ContextPath, "**/*"))
-	if err != nil {
-		return
-	}
-
-	res := make([]string, 0, len(sources))
-	for _, src := range sources {
-		if stat, err := os.Stat(src); err != nil {
-			return err
-		} else if stat.IsDir() {
-			res = append(res, strings.TrimPrefix(src, p.ContextPath))
-			continue
-		}
-
-		file, err := os.OpenFile(src, os.O_RDONLY, 0644)
-		if err != nil {
-			return err
-		}
-
-		hash, err := highwayhash.New(hashKey)
-		if err != nil {
-			file.Close()
-			return err
-		}
-
-		_, err = io.Copy(hash, file)
-		if err != nil {
-			file.Close()
-			return err
-		}
-
-		err = file.Close()
-		if err != nil {
-			return err
-		}
-
-		res = append(res, fmt.Sprintf("%s:%s", strings.TrimPrefix(src, p.ContextPath), hex.EncodeToString(hash.Sum(nil))))
-	}
-
-	if baseref != "" {
-		fmt.Fprintf(out, "Baseref: %s\n", baseref)
-	}
-	fmt.Fprintf(out, "Dockerfile: %s\n", string(p.Dockerfile))
-	fmt.Fprintf(out, "Sources:\n%s\n", strings.Join(res, "\n"))
-	return nil
-}
-
 // BaseRef returns the ref of the base image of a project
 func (p *Project) BaseRef(build reference.Named) (reference.NamedTagged, error) {
 	hash, err := p.Base.hash("")
@@ -616,11 +433,6 @@ func (p *Project) BaseRef(build reference.Named) (reference.NamedTagged, error) 
 		return nil, err
 	}
 	return reference.WithTag(build, fmt.Sprintf("base--%s", hash))
-}
-
-// ChunkRef returns the ref of a chunk image
-func (p *Project) ChunkRef(build reference.Named, chunk string) (reference.NamedTagged, error) {
-	return reference.WithTag(build, chunk)
 }
 
 func (p *ProjectChunk) buildAsBase(ctx context.Context, dest reference.Named, sess *BuildSession) (absref reference.Digested, err error) {
@@ -754,37 +566,6 @@ func (p *ProjectChunk) build(ctx context.Context, sess *BuildSession) (chkRef re
 	sess.recordChunk(chkRef.String(), mf)
 
 	return
-}
-
-// ChunkImageType describes the chunk build artifact type
-type ChunkImageType string
-
-const (
-	// ImageTypeTest is an image built for testing
-	ImageTypeTest ChunkImageType = "test"
-	// ImageTypeFull is the full chunk image
-	ImageTypeFull ChunkImageType = "full"
-	// ImageTypeChunked is the chunk image with the base layers removed
-	ImageTypeChunked ChunkImageType = "chunked"
-	// ImageTypeChunkedNoHash is the chunk image with the base layers removed and no hash in the name
-	ImageTypeChunkedNoHash ChunkImageType = "chunked-wohash"
-)
-
-// ImageName produces a chunk image name
-func (p *ProjectChunk) ImageName(tpe ChunkImageType, sess *BuildSession) (reference.NamedTagged, error) {
-	if sess.baseRef == nil {
-		return nil, fmt.Errorf("base ref not set")
-	}
-
-	if tpe == ImageTypeChunkedNoHash {
-		return reference.WithTag(sess.Dest, fmt.Sprintf("%s", p.Name))
-	}
-
-	hash, err := p.hash(sess.baseRef.String())
-	if err != nil {
-		return nil, fmt.Errorf("cannot compute chunk hash: %w", err)
-	}
-	return reference.WithTag(sess.Dest, fmt.Sprintf("%s--%s--%s", p.Name, hash, tpe))
 }
 
 func (p *ProjectChunk) buildImage(ctx context.Context, tpe ChunkImageType, sess *BuildSession) (tgt reference.Named, didBuild bool, err error) {
